Refactor form module
Form module was supposed to be universal for any project where it will
be used. Previous form didn't met that requirement. On top of that,
some of functionalities for form wasn't event part of form class,
which deny principles of object oriented programming. From now on form
class can be used in any project and has methods that should be in
it from the begining f.e: getValues to get values from form inputs.

As project will grow, more funcionalities will be added to Form when
needed.